feat: working additional brands
