+++ src/wizards/newspack/views/settings/additional-brands/types.d.ts
@@ -1,3 +1,10 @@
+interface ThemeColorsMeta {
+	color: string;
+	name: string;
+	theme_mod_name?: string;
+	default?: string;
+}
+
 type Brand = {
 	id: number;
 	count: number;
@@ -10,12 +17,8 @@ type Brand = {
 	meta: {
 		_custom_url: string;
 		_show_page_on_front: number;
-		_logo: number;
-		_theme_colors: {
-			color: string;
-			name: string;
-			theme_mod_name: string;
-		}[];
+		_logo: number | Attachment;
+		_theme_colors: ThemeColorsMeta[];
 		_menus: Array< {
 			location: string;
 			menu: number;
